% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cars}
\alias{cars}
\title{Speed and Stopping Distances of Cars}
\format{
\subsection{\code{cars}}{

A data frame with 50 observations on 2 variables.
\describe{
\item{speed}{numeric Speed (mph)}
\item{dist}{numeric Stopping distance (ft)}
}
}
}
\source{
Ezekiel, M. (1930) \emph{Methods of Correlation Analysis}. Wiley.
}
\usage{
cars
}
\description{
The data provide the speed and the distances to stop of cars which were recorded in 1920s.
}
\examples{
plot(cars, xlab = "Speed (mph)", ylab = "Stopping distance (ft)")
}
\references{
McNeil, D. R. (1977) \emph{Interactive Data Analysis}. Wiley.
}
\keyword{datasets}
