% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{mymean}
\alias{mymean}
\title{Find the mean value of a numeric vector}
\usage{
mymean(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
Mean value of the input vector.
}
\description{
Find the mean value of a numeric vector
}
\examples{
mymean(1:10)
mymean(c(2, 3, 4, 5))
}
